Restructure repo
All React files are now found in src/ except for App.js

Files for compiling the Android app are in the android folder
