+++ android/.gitignore
@@ -1,0 +1,14 @@
+# Ignore Gradle project-specific cache directory
+.gradle
+
+# Ignore Gradle build output directory
+build
+
+# # Not sure if we should ignore the Gradle Wrapper files...
+# gradlew
+# gradlew.bat
+
+# Ignore keystores
+*.keystore
+
+.idea
