+++ android/app/build.gradle
@@ -1,0 +1,28 @@
+/*
+ * This file was generated by the Gradle 'init' task.
+ *
+ * This generated file contains a sample Java application project to get you started.
+ * For more details take a look at the 'Building Java & JVM projects' chapter in the Gradle
+ * User Manual available at https://docs.gradle.org/7.1.1/userguide/building_java_projects.html
+ */
+
+apply plugin: 'com.android.application'
+
+android {
+	compileSdkVersion 29
+	buildToolsVersion "29.0.2"
+
+//	defaultConfig {
+//		applicationId "com.soumya.sethy.reactnativesetupandroid"
+//	}
+	defaultConfig {
+		applicationId "com.projectrainbow"
+		minSdkVersion 16
+		targetSdkVersion 29
+		resConfigs "en"
+	}
+}
+
+dependencies {
+	implementation "com.facebook.react:react-native:+"
+}
