+++ android/app/src/main/AndroidManifest.xml
@@ -1,0 +1,25 @@
+<?xml version="1.0" encoding="utf-8"?>
+<manifest xmlns:android="http://schemas.android.com/apk/res/android"
+    package="com.projectrainbow"
+    android:versionCode="1"
+    android:versionName="1.0" >
+
+<uses-permission android:name="android.permission.INTERNET"/>
+
+<application
+    android:allowBackup="true"
+    android:supportsRtl="true"
+    >
+
+    <activity
+        android:name=".activity.ChooseAreaActivity"
+    >
+        <intent-filter>
+            <action android:name="android.intent.action.MAIN" />
+
+            <category android:name="android.intent.category.LAUNCHER" />
+        </intent-filter>
+    </activity>
+</application>
+
+</manifest>
