+++ src/assets/images/flower.png
