+++ src/components/AdjustableText.js
