+++ src/components/ChannelCollection.js
