+++ src/components/RainbowChannel.js
