+++ src/components/SearchArea.js
