+++ src/scripts/GetProps.js
